refactor(#3281): extract built-in global ctor dispatch from compileNewExpression (WAVE-C slice 1)
Lift the built-in global constructor dispatch band (~1,096 LOC) VERBATIM out of
`compileNewExpression` (src/codegen/expressions/new-super.ts) into a new sibling
module `src/codegen/expressions/new-builtin-globals.ts`
(`tryCompileBuiltinGlobalNew`).

Covered arms: Promise, the Number/String/Boolean wrapper objects, the Error
family (Error/TypeError/RangeError/SyntaxError/URIError/EvalError/ReferenceError/
Test262Error), AggregateError, SuppressedError, Object, Proxy, Function, Date,
and the TypedArray constructors.

Byte-identity sentinel lift: each arm's internal `return`s (including nested
arrow-closure returns) are left verbatim; a single `NEW_GLOBAL_FALLTHROUGH`
sentinel signals "not one of these globals" so the caller resumes its dispatch
ladder. Four helpers used by the band (isStringTypedArg,
emitHostTaBufferConstruct, hostTaBufferArgSymName, resolvesToAmbientGlobal) are
now `export`ed from new-super.ts.

- compileNewExpression: ~3,082 -> ~1,994 LOC
- new-builtin-globals.ts: 1,155 LOC (< 1,500 threshold)
- prove-emit-identity: IDENTICAL across all 39 gc/standalone/wasi emits
- tsc --noEmit clean; tests/issue-3281.test.ts (11 cases) green

Slice 1 of 2 (#3281, subtask of #3182). Slice 2 extracts the indexed builtins
(ArrayBuffer/DataView/Array) and brings compileNewExpression under 1,500 LOC.

✓
