+++ src/codegen/expressions/new-builtin-globals.ts
@@ -1,0 +1,1155 @@
+import type { Instr, ValType } from "../../ir/types.js";
+// Copyright (c) 2026 Loopdive GmbH. Licensed under Apache-2.0 WITH LLVM-exception.
+/**
+ * Built-in global constructor dispatch for `new C(...)` — extracted from
+ * `compileNewExpression` (new-super.ts) as WAVE-C decomposition slice 1 (#3281).
+ *
+ * Covers the identifier-keyed built-in globals: Promise, the
+ * Number/String/Boolean wrapper objects, the Error family
+ * (Error/TypeError/RangeError/…), AggregateError, SuppressedError, Object,
+ * Proxy, Function, Date, and the TypedArray constructors. Each arm either fully
+ * handles the ctor (emitting into `fctx.body` and returning a `ValType | null`)
+ * or does nothing and falls through; a single `NEW_GLOBAL_FALLTHROUGH` sentinel
+ * signals "not one of these globals" so the caller resumes its dispatch ladder.
+ * The lifted arms are byte-identical to the inline originals.
+ */
+import { ts } from "../../ts-api.js";
+import type { CodegenContext, FunctionContext } from "../context/types.js";
+import { allocLocal, allocTempLocal, releaseTempLocal } from "../context/locals.js";
+import { addUnionImports, getArrTypeIdxFromVec, getOrRegisterVecType, typedArrayVecStorage } from "../index.js";
+import { coercionPlan } from "../coercion-plan.js";
+import { emitTaViewConstruct, emitTaViewConstructWindowed } from "../dataview-native.js";
+import { emitNativeDateParse } from "../date-parse-native.js";
+import { compileObjectLiteralAsExternref } from "../literals.js";
+import { ensureAnyToStringHelper } from "../native-strings.js";
+import { emitNativeNumberFormat } from "../number-format-native.js";
+import { ensureObjectRuntime } from "../object-runtime.js";
+import { emitStandalonePromiseFromExecutor, emitStandalonePromiseFromExecutorValue } from "../promise-executor.js";
+import { emitStandaloneTest262Error, emitWasiErrorConstructor, isWasiErrorName } from "../registry/error-types.js";
+import { coerceType, compileExpression } from "../shared.js";
+import { compileStringLiteral } from "../string-ops.js";
+import { coerceType as coerceTypeImpl } from "../type-coercion.js";
+import { ensureDateDaysFromCivilHelper, ensureDateStruct } from "./builtins.js";
+import { emitObjectCoercion } from "./calls-guards.js";
+import {
+  emitDynamicNewFunctionHostEval,
+  emitStandaloneDynamicFunctionStub,
+  isGlobalFunctionIdentifier,
+  tryStaticNewFunction,
+} from "./eval-inline.js";
+import { emitThrowTypeError, noJsHost } from "./helpers.js";
+import { ensureLateImport, flushLateImportShifts } from "./late-imports.js";
+import {
+  emitHostTaBufferConstruct,
+  hostTaBufferArgSymName,
+  isStringTypedArg,
+  resolvesToAmbientGlobal,
+} from "./new-super.js";
+
+/** Sentinel: the `new` target is not one of the built-in global constructors. */
+export const NEW_GLOBAL_FALLTHROUGH = Symbol("new-builtin-global-fallthrough");
+
+export function tryCompileBuiltinGlobalNew(
+  ctx: CodegenContext,
+  fctx: FunctionContext,
+  expr: ts.NewExpression,
+): ValType | null | typeof NEW_GLOBAL_FALLTHROUGH {
+  // Handle `new Promise(executor)`.
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Promise") {
+    // (#2959) Native standalone/WASI path — construct the `$Promise` and run the
+    // executor with synthesised native resolve/reject closures, retiring the
+    // `Promise_new` host import. Gated inside the helper on
+    // `isStandalonePromiseActive` + a resolvable executor closure; when it can't
+    // apply it emits NOTHING and returns false, falling through to the host path
+    // below (byte-unchanged in host/gc mode). Guard the ambient-global binding so
+    // a user `class Promise {}` / local shadow keeps the normal ctor path.
+    const promiseArgs = expr.arguments ?? [];
+    if (
+      promiseArgs.length >= 1 &&
+      !ctx.classSet.has("Promise") &&
+      resolvesToAmbientGlobal(ctx, expr.expression) &&
+      emitStandalonePromiseFromExecutor(ctx, fctx, promiseArgs[0]!)
+    ) {
+      return { kind: "externref" };
+    }
+    // (#2903 R1) Native path for a NON-inline executor VALUE (identifier / param
+    // / any runtime closure) — invoked through `__apply_closure` since its
+    // concrete closure type isn't recoverable at compile time. Tried only when
+    // the inline path above declined AND the executor arg is not a syntactic
+    // arrow/function-expression (those are the inline path's domain). Retires the
+    // `Promise_new` + `__make_callback` leak for `make(ex)=>new Promise(ex)`.
+    if (
+      promiseArgs.length >= 1 &&
+      !ctx.classSet.has("Promise") &&
+      resolvesToAmbientGlobal(ctx, expr.expression) &&
+      !ts.isArrowFunction(promiseArgs[0]!) &&
+      !ts.isFunctionExpression(promiseArgs[0]!) &&
+      emitStandalonePromiseFromExecutorValue(ctx, fctx, () => {
+        const t = compileExpression(ctx, fctx, promiseArgs[0]!, { kind: "externref" });
+        if (t && t.kind !== "externref") coerceType(ctx, fctx, t, { kind: "externref" });
+        else if (t === null) fctx.body.push({ op: "ref.null.extern" });
+      })
+    ) {
+      return { kind: "externref" };
+    }
+    // (#2903) This is the genuine HOST `new Promise` fallthrough (executor not
+    // native-lowerable) — the runtime value is a host promise, so the
+    // `.then`/`.catch` bridge's miss arm must keep its host fallback for the
+    // rest of this module. (`Promise_new` funcMap presence can't signal this:
+    // it is upfront-registered for every syntactic `new Promise` even when the
+    // lowering is native.) Order caveat: a bridge compiled BEFORE this point
+    // already chose its arm; acceptable — the affected shape (a `.then` in an
+    // earlier function over a later non-inline-executor promise) now throws a
+    // catchable TypeError instead of host-chaining, and can only occur in
+    // modules that were irreducibly host-import-leaky anyway.
+    if (ctx.standalone === true && ctx.wasi !== true) {
+      ctx.moduleHasHostPromiseSource = true;
+    }
+    let funcIdx =
+      ctx.funcMap.get("Promise_new") ??
+      ensureLateImport(ctx, "Promise_new", [{ kind: "externref" }], [{ kind: "externref" }]);
+    flushLateImportShifts(ctx, fctx);
+    funcIdx = ctx.funcMap.get("Promise_new") ?? funcIdx;
+    if (funcIdx !== undefined) {
+      const args = expr.arguments ?? [];
+      if (args.length >= 1) {
+        compileExpression(ctx, fctx, args[0]!, { kind: "externref" });
+      } else {
+        fctx.body.push({ op: "ref.null.extern" });
+      }
+      fctx.body.push({ op: "call", funcIdx });
+    } else {
+      fctx.body.push({ op: "ref.null.extern" });
+    }
+    return { kind: "externref" };
+  }
+
+  // Handle `new Number(x)`, `new String(x)`, `new Boolean(x)` — wrapper constructors
+  // Return externref so typeof returns "object" (wrapper semantics).
+  // Number/Boolean: box to externref via __box_number. String: already externref.
+  if (ts.isIdentifier(expr.expression)) {
+    const ctorName = expr.expression.text;
+    if (ctorName === "Number" || ctorName === "String" || ctorName === "Boolean") {
+      const args = expr.arguments ?? [];
+
+      if (ctorName === "Number") {
+        // new Number(x) → create real JS Number wrapper object via __new_Number host import
+        // (typeof new Number(0) === "object", not "number")
+        if (args.length >= 1) {
+          // ToNumber(Symbol) throws TypeError (§7.1.4) — the wrapper ctor runs
+          // ToNumber on its argument before boxing. Mirror the `Number(sym)`
+          // call-path guard so `new Number(Symbol())` throws too (#1564).
+          if (ctx.oracle.staticJsTypeOf(args[0]!) === "symbol") {
+            const t = compileExpression(ctx, fctx, args[0]!);
+            if (t !== null) fctx.body.push({ op: "drop" });
+            emitThrowTypeError(ctx, fctx, "Cannot convert a Symbol value to a number");
+            return { kind: "externref" };
+          }
+          compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+        } else {
+          fctx.body.push({ op: "f64.const", value: 0 });
+        }
+        const newNumIdx = ensureLateImport(ctx, "__new_Number", [{ kind: "f64" }], [{ kind: "externref" }]);
+        flushLateImportShifts(ctx, fctx);
+        const finalNumIdx = ctx.funcMap.get("__new_Number") ?? newNumIdx;
+        if (finalNumIdx !== undefined) fctx.body.push({ op: "call", funcIdx: finalNumIdx });
+        return { kind: "externref" };
+      }
+
+      if (ctorName === "String") {
+        // new String(x) → create real JS String wrapper object via __new_String host import
+        // (typeof new String("") === "object", not "string")
+        if (args.length >= 1) {
+          compileExpression(ctx, fctx, args[0]!, { kind: "externref" });
+        } else {
+          const emptyStrResult = compileStringLiteral(ctx, fctx, "");
+          if (!emptyStrResult) {
+            fctx.body.push({ op: "ref.null.extern" });
+          }
+        }
+        const newStrIdx = ensureLateImport(ctx, "__new_String", [{ kind: "externref" }], [{ kind: "externref" }]);
+        flushLateImportShifts(ctx, fctx);
+        const finalStrIdx = ctx.funcMap.get("__new_String") ?? newStrIdx;
+        if (finalStrIdx !== undefined) fctx.body.push({ op: "call", funcIdx: finalStrIdx });
+        return { kind: "externref" };
+      }
+
+      if (ctorName === "Boolean") {
+        // new Boolean(x) → create real JS Boolean wrapper object via __new_Boolean host import
+        // (typeof new Boolean(false) === "object", not "boolean")
+        if (args.length >= 1) {
+          // ToBoolean never throws on Symbol (a Symbol is truthy), but this path
+          // coerces the arg to f64 first, which would silently lose the Symbol.
+          // A Symbol arg should produce a truthy wrapper: box 1.0.
+          if (ctx.oracle.staticJsTypeOf(args[0]!) === "symbol") {
+            const t = compileExpression(ctx, fctx, args[0]!);
+            if (t !== null) fctx.body.push({ op: "drop" });
+            fctx.body.push({ op: "f64.const", value: 1 });
+          } else {
+            compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+          }
+        } else {
+          fctx.body.push({ op: "f64.const", value: 0 });
+        }
+        const newBoolIdx = ensureLateImport(ctx, "__new_Boolean", [{ kind: "f64" }], [{ kind: "externref" }]);
+        flushLateImportShifts(ctx, fctx);
+        const finalBoolIdx = ctx.funcMap.get("__new_Boolean") ?? newBoolIdx;
+        if (finalBoolIdx !== undefined) fctx.body.push({ op: "call", funcIdx: finalBoolIdx });
+        return { kind: "externref" };
+      }
+    }
+  }
+
+  // Handle `new Error(msg)`, `new TypeError(msg)`, `new RangeError(msg)` — create real Error objects
+  // via host import so .name, .message, .stack are correct and instanceof works.
+  // Standalone fallback: the thrown value is just the message string (as before).
+  if (ts.isIdentifier(expr.expression)) {
+    const ctorName = expr.expression.text;
+    if (
+      ctorName === "Error" ||
+      ctorName === "TypeError" ||
+      ctorName === "RangeError" ||
+      ctorName === "SyntaxError" ||
+      ctorName === "URIError" ||
+      ctorName === "EvalError" ||
+      ctorName === "ReferenceError" ||
+      ctorName === "Test262Error"
+    ) {
+      const args = expr.arguments ?? [];
+      if (args.length >= 1) {
+        // Compile the message argument to externref
+        const resultType = compileExpression(ctx, fctx, args[0]!, {
+          kind: "externref",
+        });
+        if (resultType && resultType.kind !== "externref") {
+          coerceType(ctx, fctx, resultType, { kind: "externref" });
+        }
+      } else {
+        // No message — push null externref (undefined message)
+        fctx.body.push({ op: "ref.null.extern" });
+      }
+      // (#2969) §20.5.1.1 step 3 — `msg = ToString(message)` at CONSTRUCTION
+      // time. In standalone/WASI the native `__new_<Name>` ctor stores its arg
+      // verbatim (see error-types.ts), so `new Error(42).message` was the number
+      // `42` (spec: `"42"`) and `String(new Error(42))` degraded to `"Error"`.
+      // Do the ToString HERE at the user call site (null-guarded so argument-less
+      // / `new Error(undefined)` still render the name alone) rather than inside
+      // the SHARED ctor — the ctor is also lazily emitted for internal compiler
+      // error paths (destructuring/coercion `TypeError`s), and pulling the
+      // `__any_to_string` family into those emissions destabilised standalone
+      // `any`-equality dispatch (the tag-5 value-eq gap deferred to #2580 M2 /
+      // #3032). Host mode's `__new_<Name>` import does ToString in JS, so only
+      // the native path needs this. Applies to both the WASI-error and
+      // Test262Error branches below (both native, both standalone/WASI).
+      if ((ctx.wasi || ctx.standalone) && args.length >= 1) {
+        // Force `number_toString` before `__any_to_string` bakes so its number
+        // arm renders a raw numeric message ("42") instead of degrading to
+        // "[object Object]" (a module that only constructs `new Error(n)` never
+        // otherwise pulls the number formatter). Must precede the ensure below.
+        if (ctx.funcMap.get("number_toString") === undefined) {
+          emitNativeNumberFormat(ctx, new Set(["number_toString"]));
+        }
+        const anyToStrIdx = ensureAnyToStringHelper(ctx);
+        if (anyToStrIdx >= 0) {
+          const msgTmp = allocTempLocal(fctx, { kind: "externref" });
+          fctx.body.push(
+            { op: "local.set", index: msgTmp },
+            { op: "local.get", index: msgTmp },
+            { op: "ref.is_null" },
+            {
+              op: "if",
+              blockType: { kind: "val", type: { kind: "externref" } },
+              // undefined / null argument → keep null (renders name alone).
+              then: [{ op: "ref.null.extern" }],
+              // ToString(message): externref → anyref → __any_to_string
+              // (ref $AnyString) → externref for the ctor's $message field.
+              else: [
+                { op: "local.get", index: msgTmp },
+                { op: "any.convert_extern" },
+                { op: "call", funcIdx: anyToStrIdx },
+                { op: "extern.convert_any" },
+              ],
+            },
+          );
+          releaseTempLocal(fctx, msgTmp);
+        }
+      }
+      // (#1104 Phase 1) In WASI/standalone mode, the JS host is unavailable —
+      // use a Wasm-native `__new_<Name>` function that builds a `$Error_struct`
+      // instead of a `env.__new_<Name>` host import that would leave the
+      // module unsatisfiable at instantiation time. JS-host mode is unchanged.
+      const importName = `__new_${ctorName}`;
+      // #1473 — standalone mode also has no JS host; build the Error in-module.
+      if ((ctx.wasi || ctx.standalone) && isWasiErrorName(ctorName)) {
+        emitWasiErrorConstructor(ctx, ctorName, 1);
+        const internalFuncIdx = ctx.funcMap.get(importName);
+        if (internalFuncIdx !== undefined) {
+          fctx.body.push({ op: "call", funcIdx: internalFuncIdx });
+        }
+        return { kind: "externref" };
+      }
+      // (#2902) `Test262Error` is not a WASI builtin error, but the test262
+      // harness declares it (`class Test262Error extends Error`) and `throw new
+      // Test262Error(...)` appears in nearly every wrapped test. In standalone /
+      // WASI mode the `__new_Test262Error` host import is unsatisfiable and
+      // leaks the module out of host-free — yet the ctor is only reached on the
+      // untaken failure path of a passing test. Build it natively as an
+      // $Error_struct (tagged Error, name "Test262Error") so ~2,779 such tests
+      // become host-free. JS-host mode keeps the host import (real Error).
+      if ((ctx.wasi || ctx.standalone) && ctorName === "Test262Error") {
+        emitStandaloneTest262Error(ctx, 1);
+        const internalFuncIdx = ctx.funcMap.get(importName);
+        if (internalFuncIdx !== undefined) {
+          fctx.body.push({ op: "call", funcIdx: internalFuncIdx });
+        }
+        return { kind: "externref" };
+      }
+      // Use host import to create a real Error object with correct .name/.message/.stack
+      const funcIdx = ensureLateImport(
+        ctx,
+        importName,
+        [{ kind: "externref" }], // message param
+        [{ kind: "externref" }], // returns Error object
+      );
+      flushLateImportShifts(ctx, fctx);
+      if (funcIdx !== undefined) {
+        fctx.body.push({ op: "call", funcIdx });
+      }
+      // If import not available (standalone), value is already on stack as externref message
+      return { kind: "externref" };
+    }
+  }
+
+  // Handle `new AggregateError(errors, message, options?)` (#844)
+  // AggregateError takes (iterable, message, options?) — pass errors and message as externref
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "AggregateError") {
+    const args = expr.arguments ?? [];
+    // Compile errors argument (iterable) as externref
+    if (args.length >= 1) {
+      const errorsType = compileExpression(ctx, fctx, args[0]!, {
+        kind: "externref",
+      });
+      if (errorsType && errorsType.kind !== "externref") {
+        coerceType(ctx, fctx, errorsType, { kind: "externref" });
+      }
+    } else {
+      fctx.body.push({ op: "ref.null.extern" });
+    }
+    // Compile message argument as externref
+    if (args.length >= 2) {
+      const msgType = compileExpression(ctx, fctx, args[1]!, {
+        kind: "externref",
+      });
+      if (msgType && msgType.kind !== "externref") {
+        coerceType(ctx, fctx, msgType, { kind: "externref" });
+      }
+    } else {
+      fctx.body.push({ op: "ref.null.extern" });
+    }
+    // Compile options argument as externref (for cause property)
+    if (args.length >= 3) {
+      const optsType = compileExpression(ctx, fctx, args[2]!, {
+        kind: "externref",
+      });
+      if (optsType && optsType.kind !== "externref") {
+        coerceType(ctx, fctx, optsType, { kind: "externref" });
+      }
+    } else {
+      fctx.body.push({ op: "ref.null.extern" });
+    }
+    const funcIdx = ensureLateImport(
+      ctx,
+      "__new_AggregateError",
+      [{ kind: "externref" }, { kind: "externref" }, { kind: "externref" }],
+      [{ kind: "externref" }],
+    );
+    flushLateImportShifts(ctx, fctx);
+    if (funcIdx !== undefined) {
+      fctx.body.push({ op: "call", funcIdx });
+    }
+    return { kind: "externref" };
+  }
+
+  // Handle `new SuppressedError(error, suppressed, message, options?)` (#1634).
+  // Spec §20.5.10.1: all four arguments are externref; `options.cause` is
+  // installed via the dedicated `__new_SuppressedError` host import. The generic
+  // 3-param extern-class path dropped `options` (no `cause`) and mishandled the
+  // message coercion, so route through the dedicated import like AggregateError.
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "SuppressedError") {
+    const args = expr.arguments ?? [];
+    for (let i = 0; i < 4; i++) {
+      if (args.length > i) {
+        const t = compileExpression(ctx, fctx, args[i]!, { kind: "externref" });
+        if (t && t.kind !== "externref") {
+          coerceType(ctx, fctx, t, { kind: "externref" });
+        }
+      } else {
+        fctx.body.push({ op: "ref.null.extern" });
+      }
+    }
+    const funcIdx = ensureLateImport(
+      ctx,
+      "__new_SuppressedError",
+      [{ kind: "externref" }, { kind: "externref" }, { kind: "externref" }, { kind: "externref" }],
+      [{ kind: "externref" }],
+    );
+    flushLateImportShifts(ctx, fctx);
+    if (funcIdx !== undefined) {
+      fctx.body.push({ op: "call", funcIdx });
+    }
+    return { kind: "externref" };
+  }
+
+  // Handle `new Object()` — create an empty object (equivalent to `{}`).
+  // (#1343) Previously this emitted `ref.null.extern`, but JS spec treats
+  // `new Object()` as a real object: `Boolean(new Object()) === true`,
+  // `(new Object()).hasOwnProperty(...) === false`, etc. Returning null
+  // externref made the receiver fall through every host-import branch
+  // expecting a real object, e.g. `Boolean(new Object())` returned `false`
+  // because `__to_boolean(null) === 0`.
+  //
+  // Use `__new_plain_object` host import to produce a fresh empty object
+  // with the ordinary `Object.prototype` prototype (#1525). `new Object()`
+  // per §20.1.1.1 must inherit `Object.prototype` — using `__object_create(null)`
+  // gave it a null prototype, so it had no `toString`/`valueOf` and any
+  // ToPrimitive coercion (`==`, arithmetic, `String(...)`) threw
+  // "Cannot convert object to primitive value" instead of producing
+  // "[object Object]". Falls back to `ref.null.extern` only if the import
+  // can't be registered.
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Object") {
+    // (#3118) `new Object(v)` is spec-identical to `Object(v)` (§20.1.1.1:
+    // return ToObject(v)). This arm previously ignored its arg and built an
+    // empty object; delegate to the shared coercion so a primitive boxes to its
+    // wrapper (an object passes through, null/undefined/none → fresh object).
+    const r = emitObjectCoercion(ctx, fctx, expr.arguments ?? []);
+    return r === null ? { kind: "externref" } : (r as ValType);
+  }
+
+  // Handle `new Proxy(target, handler)`.
+  //
+  // JS-host mode: delegate to the `__proxy_create(target, handler)` host import
+  // (the host wraps the target in a real JS Proxy with the given handler).
+  //
+  // Standalone mode (#1100 Phase 1): there is no host Proxy, so route through the
+  // Wasm-native `__proxy_create(target, handler)` emitted by `ensureObjectRuntime`
+  // (object-runtime.ts `ensureProxyRuntime`). It reads the get/set/has/apply trap
+  // closures off the handler object at runtime, allocates a `$Proxy` (subtype of
+  // `$Object`), and the property-runtime front-guards (`__extern_get/set/has`)
+  // dispatch reads/writes/has through the traps. Both modes share the same
+  // `(target, handler) -> externref` signature, so the call site is uniform.
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Proxy") {
+    if (ctx.standalone) {
+      const args = expr.arguments ?? [];
+      // Force the object runtime (which registers the native __proxy_create +
+      // the trap dispatch helpers + the front-guards) before we look up the idx.
+      ensureObjectRuntime(ctx);
+      const compileToExternref = (arg: ts.Expression | undefined): void => {
+        if (arg === undefined) {
+          fctx.body.push({ op: "ref.null.extern" });
+          return;
+        }
+        // An OBJECT-LITERAL handler/target must lower to an OPEN `$Object`
+        // (`__new_plain_object` + `__extern_set` per prop) so the runtime
+        // `__proxy_create` can read the traps off the handler via `__extern_get`.
+        // A closed typed struct (the default for an inline literal) hides its
+        // fields from the open-object prop-map walk, so every trap reads null and
+        // never fires. `compileObjectLiteralAsExternref` builds the open form —
+        // the same shape a `const h: any = {…}` handler takes.
+        if (ts.isObjectLiteralExpression(arg)) {
+          const r = compileObjectLiteralAsExternref(ctx, fctx, arg);
+          if (r === null) {
+            // Builder unavailable — push undefined so the body stays valid.
+            fctx.body.push({ op: "ref.null.extern" });
+          }
+          return;
+        }
+        const r = compileExpression(ctx, fctx, arg, { kind: "externref" });
+        if (r && r.kind !== "externref") {
+          if (r.kind === "ref" || r.kind === "ref_null") {
+            fctx.body.push({ op: "extern.convert_any" });
+          } else {
+            coerceTypeImpl(ctx, fctx, r, { kind: "externref" });
+          }
+        } else if (!r) {
+          // void result (shouldn't happen for a value arg) — push undefined.
+          fctx.body.push({ op: "ref.null.extern" });
+        }
+      };
+      compileToExternref(args[0]);
+      compileToExternref(args[1]);
+      const proxyCreateIdx = ctx.funcMap.get("__proxy_create");
+      if (proxyCreateIdx !== undefined) {
+        fctx.body.push({ op: "call", funcIdx: proxyCreateIdx });
+      } else {
+        // Runtime not available (should not happen) — drop args, push undefined.
+        fctx.body.push({ op: "drop" });
+        fctx.body.push({ op: "drop" });
+        fctx.body.push({ op: "ref.null.extern" });
+      }
+      return { kind: "externref" };
+    }
+    const args = expr.arguments ?? [];
+    if (args.length >= 1) {
+      // Compile target argument and coerce to externref
+      const bodyBefore = fctx.body.length;
+      const targetResult = compileExpression(ctx, fctx, args[0]!);
+      if (targetResult && targetResult.kind !== "externref") {
+        if (targetResult.kind === "ref" || targetResult.kind === "ref_null") {
+          fctx.body.push({ op: "extern.convert_any" });
+        } else {
+          coerceTypeImpl(ctx, fctx, targetResult, { kind: "externref" });
+        }
+      }
+
+      // Compile handler argument and coerce to externref (or push null if missing)
+      if (args.length >= 2) {
+        const handlerResult = compileExpression(ctx, fctx, args[1]!);
+        if (handlerResult && handlerResult.kind !== "externref") {
+          if (handlerResult.kind === "ref" || handlerResult.kind === "ref_null") {
+            fctx.body.push({ op: "extern.convert_any" });
+          } else {
+            coerceTypeImpl(ctx, fctx, handlerResult, { kind: "externref" });
+          }
+        }
+      } else {
+        fctx.body.push({ op: "ref.null.extern" });
+      }
+
+      // Emit call to __proxy_create(target, handler) -> externref
+      const proxyIdx = ensureLateImport(
+        ctx,
+        "__proxy_create",
+        [{ kind: "externref" }, { kind: "externref" }],
+        [{ kind: "externref" }],
+      );
+      flushLateImportShifts(ctx, fctx);
+      if (proxyIdx !== undefined) {
+        fctx.body.push({ op: "call", funcIdx: proxyIdx });
+      }
+
+      return { kind: "externref" };
+    }
+    // No arguments — `new Proxy()`. Per §28.2.1.1 the missing target/handler
+    // are `undefined`, which are not objects, so construction throws TypeError.
+    // Route through __proxy_create(null, null) so the runtime raises it (#2180).
+    {
+      fctx.body.push({ op: "ref.null.extern" });
+      fctx.body.push({ op: "ref.null.extern" });
+      const proxyIdx = ensureLateImport(
+        ctx,
+        "__proxy_create",
+        [{ kind: "externref" }, { kind: "externref" }],
+        [{ kind: "externref" }],
+      );
+      flushLateImportShifts(ctx, fctx);
+      if (proxyIdx !== undefined) {
+        fctx.body.push({ op: "call", funcIdx: proxyIdx });
+      }
+    }
+    return { kind: "externref" };
+  }
+
+  // Handle `new Function(...)`.
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Function") {
+    const args = expr.arguments ?? [];
+    // (#2924) Constant param list + body → compile-away to a real AOT callable
+    // (global scope, no capture). Dynamic bodies fall through to the no-op stub
+    // below (the Tier-2 interpreter, #2928, handles them). Guarded on the
+    // GLOBAL `Function` intrinsic (a local shadow keeps the legacy stub path).
+    const staticFn = isGlobalFunctionIdentifier(expr.expression, ctx.checker)
+      ? tryStaticNewFunction(ctx, fctx, args)
+      : undefined;
+    if (staticFn !== undefined) return staticFn;
+    // (#2960) Dynamic body (non-constant args). No longer a silent no-op stub:
+    //  - JS-host mode → route to the meta-circular runtime-eval shim
+    //    (`__extern_eval`, global scope) so the constructed function actually
+    //    works — fixes the ~119 host Function-ctor test262 cluster.
+    //  - standalone/wasi (no host) → emit a source-located warning + a callable
+    //    value that throws catchably at CALL time (construction still succeeds,
+    //    so a program that never invokes it keeps working).
+    if (isGlobalFunctionIdentifier(expr.expression, ctx.checker)) {
+      const hostEval = emitDynamicNewFunctionHostEval(ctx, fctx, args);
+      if (hostEval !== undefined) return hostEval;
+      if (noJsHost(ctx)) {
+        return emitStandaloneDynamicFunctionStub(ctx, fctx, expr, args) as ValType;
+      }
+    }
+    // Legacy fallback (e.g. a local `Function` shadow, or JS-host with
+    // nativeStrings where the shim path is unavailable): evaluate args for side
+    // effects and return the historical null-value stub.
+    for (const arg of args) {
+      const argResult = compileExpression(ctx, fctx, arg);
+      if (argResult) {
+        fctx.body.push({ op: "drop" });
+      }
+    }
+    fctx.body.push({ op: "ref.null.extern" });
+    return { kind: "externref" };
+  }
+
+  // Handle `new Date()`, `new Date(ms)`, `new Date(y, m, d, ...)` — native Date struct
+  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Date") {
+    const dateTypeIdx = ensureDateStruct(ctx);
+    const args = expr.arguments ?? [];
+
+    if (args.length === 0) {
+      // (#1483) Under --target wasi, route `new Date()` (no args) to
+      // clock_time_get via the __wasi_date_now helper (registered up front in
+      // registerWasiImports).
+      if (ctx.wasi && ctx.funcMap.has("__wasi_date_now")) {
+        fctx.body.push({
+          op: "call",
+          funcIdx: ctx.funcMap.get("__wasi_date_now")!,
+        });
+        fctx.body.push({ op: "i64.trunc_sat_f64_s" });
+        fctx.body.push({ op: "struct.new", typeIdx: dateTypeIdx });
+        return { kind: "ref", typeIdx: dateTypeIdx };
+      }
+      // (#2164) Pure standalone has no wall clock — emit the Unix epoch (0)
+      // directly instead of leaking the unsatisfiable env::__date_now host
+      // import (which made `new Date()` a hard instantiate failure standalone,
+      // breaking unrelated Date tests). See the matching Date.now() fallback in
+      // expressions/calls.ts.
+      if (ctx.standalone === true) {
+        fctx.body.push({ op: "i64.const", value: 0n });
+        fctx.body.push({ op: "struct.new", typeIdx: dateTypeIdx });
+        return { kind: "ref", typeIdx: dateTypeIdx };
+      }
+      const dateNowIdx = ensureLateImport(ctx, "__date_now", [], [{ kind: "f64" }]);
+      if (dateNowIdx !== undefined) {
+        flushLateImportShifts(ctx, fctx);
+        fctx.body.push({ op: "call", funcIdx: dateNowIdx });
+        fctx.body.push({ op: "i64.trunc_sat_f64_s" });
+      } else {
+        fctx.body.push({ op: "i64.const", value: 0n });
+      }
+      fctx.body.push({ op: "struct.new", typeIdx: dateTypeIdx });
+      return { kind: "ref", typeIdx: dateTypeIdx };
+    }
+
+    if (args.length === 1) {
+      // new Date(ms) — millisecond timestamp.
+      //
+      // (#1344) Detect NaN input and store a sentinel i64 so subsequent getter
+      // calls (getDay, getHours, getTime, …) can return NaN per spec
+      // (`new Date(NaN).getTime() → NaN`). Without this, `i64.trunc_sat_f64_s`
+      // saturates NaN to 0 and the Date silently behaves like the epoch.
+      //
+      // (#1343) TimeClip per §21.4.1.31: if !isFinite(ms) or abs(ms) > 8.64e15,
+      // return NaN. Both NaN and out-of-range get the sentinel. ±Infinity is
+      // out-of-range (abs > 8.64e15), so the single magnitude check covers it.
+      //
+      // (#2164) new Date(str) — §21.4.2.1: a String value is parsed as if by
+      // Date.parse. Route a statically-string-typed arg through the pure-Wasm
+      // __date_parse helper (yields an f64 ms, NaN on failure), then fall
+      // through the same TimeClip path below. Gated to standalone / WASI for the
+      // same reason as Date.parse (host strings + lazy helper wiring trip the
+      // late-import shift class #2043); host keeps the prior ToNumber(str)→NaN.
+      if ((ctx.standalone || ctx.wasi) && isStringTypedArg(ctx, args[0]!)) {
+        emitNativeDateParse(ctx);
+        const argType = compileExpression(ctx, fctx, args[0]!, { kind: "externref" });
+        if (argType && argType.kind !== "externref") coerceType(ctx, fctx, argType, { kind: "externref" });
+        flushLateImportShifts(ctx, fctx);
+        fctx.body.push({ op: "call", funcIdx: ctx.funcMap.get("__date_parse")! });
+      } else if (
+        !ctx.standalone &&
+        !ctx.wasi &&
+        isStringTypedArg(ctx, args[0]!) &&
+        ctx.funcMap.has("__date_parse_host")
+      ) {
+        // (#2678) HOST mode: a String arg is parsed as if by Date.parse
+        // (§21.4.2.1) — delegate to the JS `Date.parse` host import (registered
+        // up-front by collectDateParseHostImports, no #2043 late-import shift).
+        const argType = compileExpression(ctx, fctx, args[0]!, { kind: "externref" });
+        if (argType && argType.kind !== "externref") coerceType(ctx, fctx, argType, { kind: "externref" });
+        fctx.body.push({ op: "call", funcIdx: ctx.funcMap.get("__date_parse_host")! });
+      } else {
+        compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+      }
+      const msLocal = allocTempLocal(fctx, { kind: "f64" });
+      fctx.body.push({ op: "local.tee", index: msLocal });
+      // isInvalid = (ms != ms) || (abs(ms) > 8.64e15)
+      // ms != ms is true iff ms is NaN (covers NaN)
+      fctx.body.push({ op: "local.get", index: msLocal });
+      fctx.body.push({ op: "f64.ne" });
+      fctx.body.push({ op: "local.get", index: msLocal });
+      fctx.body.push({ op: "f64.abs" });
+      fctx.body.push({ op: "f64.const", value: 8.64e15 });
+      fctx.body.push({ op: "f64.gt" });
+      fctx.body.push({ op: "i32.or" });
+      fctx.body.push({
+        op: "if",
+        blockType: { kind: "val", type: { kind: "i64" } },
+        then: [{ op: "i64.const", value: -9223372036854775808n }],
+        else: [{ op: "local.get", index: msLocal }, { op: "i64.trunc_sat_f64_s" }],
+      });
+      releaseTempLocal(fctx, msLocal);
+      fctx.body.push({ op: "struct.new", typeIdx: dateTypeIdx });
+      return { kind: "ref", typeIdx: dateTypeIdx };
+    }
+
+    // new Date(year, month, day?, hours?, minutes?, seconds?, ms?)
+    // JS months are 0-indexed. Day defaults to 1, rest default to 0.
+    {
+      const daysFromCivilIdx = ensureDateDaysFromCivilHelper(ctx);
+
+      // (#1343) Track whether any arg is NaN or non-finite. If so, the resulting
+      // Date is Invalid (§21.4.2.1 MakeDate / TimeClip step on non-finite).
+      // We OR-accumulate an i32 flag and stash the f64 value before trunc.
+      const nonFiniteLocal = allocTempLocal(fctx, { kind: "i32" });
+      fctx.body.push({ op: "i32.const", value: 0 });
+      fctx.body.push({ op: "local.set", index: nonFiniteLocal });
+
+      const checkNonFinite = (f64Local: number) => {
+        // flag = flag | (v != v) | (abs(v) == +Inf)
+        // We treat ±Inf as "non-finite enough" too — abs(v) > 8.64e15 is sufficient.
+        fctx.body.push({ op: "local.get", index: nonFiniteLocal });
+        fctx.body.push({ op: "local.get", index: f64Local });
+        fctx.body.push({ op: "local.get", index: f64Local });
+        fctx.body.push({ op: "f64.ne" }); // NaN check
+        fctx.body.push({ op: "i32.or" });
+        fctx.body.push({ op: "local.get", index: f64Local });
+        fctx.body.push({ op: "f64.abs" });
+        fctx.body.push({ op: "f64.const", value: 8.64e15 });
+        fctx.body.push({ op: "f64.gt" });
+        fctx.body.push({ op: "i32.or" });
+        fctx.body.push({ op: "local.set", index: nonFiniteLocal });
+      };
+
+      // Compile year
+      compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+      const yearF64Local = allocTempLocal(fctx, { kind: "f64" });
+      fctx.body.push({ op: "local.tee", index: yearF64Local });
+      checkNonFinite(yearF64Local);
+      fctx.body.push({ op: "i64.trunc_sat_f64_s" });
+      const yearLocal = allocTempLocal(fctx, { kind: "i64" });
+      fctx.body.push({ op: "local.set", index: yearLocal });
+      releaseTempLocal(fctx, yearF64Local);
+
+      // Compile month (0-indexed) + 1 for civil algorithm
+      compileExpression(ctx, fctx, args[1]!, { kind: "f64" });
+      const monthF64Local = allocTempLocal(fctx, { kind: "f64" });
+      fctx.body.push({ op: "local.tee", index: monthF64Local });
+      checkNonFinite(monthF64Local);
+      releaseTempLocal(fctx, monthF64Local);
+      fctx.body.push({ op: "i64.trunc_sat_f64_s" });
+      fctx.body.push({ op: "i64.const", value: 1n });
+      fctx.body.push({ op: "i64.add" });
+      const monthLocal = allocTempLocal(fctx, { kind: "i64" });
+      fctx.body.push({ op: "local.set", index: monthLocal });
+
+      // (#1343) For the remaining optional args, also accumulate the non-finite
+      // flag when the arg is present.
+      const compileTimePart = (argIdx: number, defaultI64: bigint, localKind: ValType) => {
+        if (args.length > argIdx) {
+          compileExpression(ctx, fctx, args[argIdx]!, { kind: "f64" });
+          const f64L = allocTempLocal(fctx, { kind: "f64" });
+          fctx.body.push({ op: "local.tee", index: f64L });
+          checkNonFinite(f64L);
+          releaseTempLocal(fctx, f64L);
+          fctx.body.push({ op: "i64.trunc_sat_f64_s" });
+        } else {
+          fctx.body.push({ op: "i64.const", value: defaultI64 });
+        }
+        const local = allocTempLocal(fctx, localKind);
+        fctx.body.push({ op: "local.set", index: local });
+        return local;
+      };
+
+      const dayLocal = compileTimePart(2, 1n, { kind: "i64" });
+      const hoursLocal = compileTimePart(3, 0n, { kind: "i64" });
+      const minutesLocal = compileTimePart(4, 0n, { kind: "i64" });
+      const secondsLocal = compileTimePart(5, 0n, { kind: "i64" });
+      const msLocal = compileTimePart(6, 0n, { kind: "i64" });
+
+      // Handle year 0-99 mapping to 1900-1999 (JS Date quirk)
+      // if (0 <= year <= 99) year += 1900
+      fctx.body.push(
+        { op: "local.get", index: yearLocal },
+        { op: "i64.const", value: 0n },
+        { op: "i64.ge_s" },
+        { op: "local.get", index: yearLocal },
+        { op: "i64.const", value: 99n },
+        { op: "i64.le_s" },
+        { op: "i32.and" },
+        {
+          op: "if",
+          blockType: { kind: "empty" },
+          then: [
+            { op: "local.get", index: yearLocal },
+            { op: "i64.const", value: 1900n },
+            { op: "i64.add" },
+            { op: "local.set", index: yearLocal },
+          ],
+        },
+      );
+
+      // Call days_from_civil(year, month, day) → i64 days
+      fctx.body.push(
+        { op: "local.get", index: yearLocal },
+        { op: "local.get", index: monthLocal },
+        { op: "local.get", index: dayLocal },
+        { op: "call", funcIdx: daysFromCivilIdx },
+      );
+
+      // timestamp = days * 86400000 + hours * 3600000 + minutes * 60000 + seconds * 1000 + ms
+      fctx.body.push(
+        { op: "i64.const", value: 86400000n },
+        { op: "i64.mul" },
+        { op: "local.get", index: hoursLocal },
+        { op: "i64.const", value: 3600000n },
+        { op: "i64.mul" },
+        { op: "i64.add" },
+        { op: "local.get", index: minutesLocal },
+        { op: "i64.const", value: 60000n },
+        { op: "i64.mul" },
+        { op: "i64.add" },
+        { op: "local.get", index: secondsLocal },
+        { op: "i64.const", value: 1000n },
+        { op: "i64.mul" },
+        { op: "i64.add" },
+        { op: "local.get", index: msLocal },
+        { op: "i64.add" },
+      );
+
+      // (#1343) TimeClip §21.4.1.31: if any arg was NaN/non-finite, or
+      // abs(ts) > 8.64e15, the time is invalid. The nonFiniteLocal flag covers
+      // the f64 NaN/Inf cases (i64.trunc_sat_f64_s would otherwise saturate them
+      // silently); the magnitude check covers in-range f64 values that still
+      // produce an out-of-range timestamp.
+      const tsResultLocal = allocTempLocal(fctx, { kind: "i64" });
+      fctx.body.push({ op: "local.set", index: tsResultLocal });
+      fctx.body.push(
+        { op: "local.get", index: nonFiniteLocal },
+        { op: "local.get", index: tsResultLocal },
+        { op: "f64.convert_i64_s" },
+        { op: "f64.abs" },
+        { op: "f64.const", value: 8.64e15 },
+        { op: "f64.gt" },
+        { op: "i32.or" },
+        {
+          op: "if",
+          blockType: { kind: "val", type: { kind: "i64" } },
+          then: [{ op: "i64.const", value: -9223372036854775808n }],
+          else: [{ op: "local.get", index: tsResultLocal }],
+        },
+      );
+      releaseTempLocal(fctx, tsResultLocal);
+      releaseTempLocal(fctx, nonFiniteLocal);
+
+      fctx.body.push({ op: "struct.new", typeIdx: dateTypeIdx });
+
+      releaseTempLocal(fctx, msLocal);
+      releaseTempLocal(fctx, secondsLocal);
+      releaseTempLocal(fctx, minutesLocal);
+      releaseTempLocal(fctx, hoursLocal);
+      releaseTempLocal(fctx, dayLocal);
+      releaseTempLocal(fctx, monthLocal);
+      releaseTempLocal(fctx, yearLocal);
+
+      return { kind: "ref", typeIdx: dateTypeIdx };
+    }
+  }
+
+  // Handle `new TypedArray(n)` — TypedArray constructors (Uint8Array, Int32Array, Float64Array, etc.)
+  // TypedArrays are fixed-length numeric arrays. Native Uint8Array uses the
+  // byte-oriented i8_byte vec; other typed arrays stay on the legacy f64
+  // representation for now.
+  if (ts.isIdentifier(expr.expression)) {
+    const TYPED_ARRAY_NAMES = new Set([
+      "Int8Array",
+      "Uint8Array",
+      "Uint8ClampedArray",
+      "Int16Array",
+      "Uint16Array",
+      "Int32Array",
+      "Uint32Array",
+      "Float32Array",
+      "Float64Array",
+    ]);
+    if (TYPED_ARRAY_NAMES.has(expr.expression.text)) {
+      // (#2593) Standalone/WASI packs integer views into i8/i16/i32 storage
+      // (Int8/Uint8/Uint8Clamped→i8_byte, Int16/Uint16→i16_byte,
+      // Int32/Uint32→i32_byte); host/gc and the float views keep f64.
+      // `typedArrayVecStorage` is the single source of truth so the
+      // count-constructor's backing vec matches the read / byteLength paths.
+      // Before #2593 only native Uint8Array packed (everything else f64), which
+      // left `new Int32Array(n)` on an f64 vec while the byteLength reader cast
+      // to i32_byte — a runtime type mismatch (read 0 / illegal cast).
+      const storage = typedArrayVecStorage(ctx, expr.expression.text);
+      const elemWasm: ValType = storage.type;
+      const elemKey = storage.key;
+      const vecTypeIdx = getOrRegisterVecType(ctx, elemKey, elemWasm);
+      const arrTypeIdx = getArrTypeIdxFromVec(ctx, vecTypeIdx);
+      const args = expr.arguments ?? [];
+
+      // (#3097) JS-host lane `new <TA>(buffer[, byteOffset[, length]])`:
+      // route through the host construct bridge (real host TypedArray view
+      // over the canonical host ArrayBuffer) instead of the numeric-length
+      // fallback below, which coerced the buffer struct to NaN → a length-0
+      // vec. Standalone keeps the native `$__ta_view` paths (B1/B2 below).
+      if (hostTaBufferArgSymName(ctx, args) !== undefined) {
+        const hostTa = emitHostTaBufferConstruct(ctx, fctx, expr.expression.text, args);
+        if (hostTa) return hostTa;
+      }
+
+      if (args.length === 0) {
+        // new TypedArray() → empty array, length 0
+        fctx.body.push({ op: "i32.const", value: 0 }); // length = 0
+        fctx.body.push({ op: "i32.const", value: 0 });
+        fctx.body.push({ op: "array.new_default", typeIdx: arrTypeIdx });
+        fctx.body.push({ op: "struct.new", typeIdx: vecTypeIdx });
+        return { kind: "ref_null", typeIdx: vecTypeIdx };
+      }
+
+      if (args.length === 1) {
+        // Check if argument is a numeric literal or expression (size constructor)
+        // vs an array/iterable (copy constructor)
+        const argType = ctx.checker.getTypeAtLocation(args[0]!);
+        const argSym = argType.getSymbol?.();
+        // #1654 — `new Uint8Array(arrayBuffer)` views the buffer's bytes. The
+        // ArrayBuffer/DataView is backed by an i32_byte vec; copy the bytes
+        // into this TypedArray's backing array. Must precede the
+        // size-constructor path (an ArrayBuffer is NOT a numeric length).
+        //
+        // #1670 — only in no-JS-host mode. The byte-buffer view path emits an
+        // unconditional `ref.cast` to the native `i32_byte` vec. In JS-host
+        // mode an ArrayBuffer / SharedArrayBuffer is NOT lowered to that vec
+        // (e.g. `new SharedArrayBuffer(n)` has no native struct), so the cast
+        // traps with `illegal cast` before any spec validation runs — this
+        // regressed 28 Atomics negative tests built on
+        // `new Int32Array(new SharedArrayBuffer(...))`. Host mode already
+        // handles the buffer arg correctly via the runtime, so skip the
+        // native view path there.
+        const argSymName = argSym?.name;
+        // (#3054 B1) Build a SHARED-BACKING `$__ta_view` that refs the buffer's
+        // vec (not a copy) so sibling views / DataViews observe writes. Gated on
+        // `noJsHost` (standalone/WASI): the native `i32_byte` vec representation
+        // of ArrayBuffer/DataView only exists in the host-free lane. In JS-host
+        // mode an ArrayBuffer is a host object (no native vec), so the recover
+        // `any.convert_extern` + `ref.cast` would trap — host mode routes buffers
+        // through the runtime instead (#1670). This replaces the former copy loop
+        // (`emitTypedArrayFromByteBuffer`) that CLONED the bytes into a fresh
+        // backing array, which broke sibling/DataView observability.
+        const taViewOk =
+          noJsHost(ctx) &&
+          (argSymName === "ArrayBuffer" || argSymName === "SharedArrayBuffer" || argSymName === "DataView");
+        if (taViewOk && !ts.isNumericLiteral(args[0]!)) {
+          const viewResult = emitTaViewConstruct(ctx, fctx, args[0]!, expr.expression.text, (e, h) =>
+            compileExpression(ctx, fctx, e, h),
+          );
+          if (viewResult) return viewResult;
+        }
+        const isArrayLike =
+          argSym?.name === "Array" ||
+          ((argType.flags & ts.TypeFlags.Object) !== 0 &&
+            argSym?.name !== undefined &&
+            TYPED_ARRAY_NAMES.has(argSym.name));
+
+        if (!isArrayLike || ts.isNumericLiteral(args[0]!)) {
+          // new TypedArray(n) → fixed-size array of length n, all zeros
+          compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+          fctx.body.push({ op: "i32.trunc_sat_f64_s" });
+          const sizeLocal = allocLocal(fctx, `__ta_size_${fctx.locals.length}`, { kind: "i32" });
+          fctx.body.push({ op: "local.tee", index: sizeLocal }); // length = n
+          fctx.body.push({ op: "local.get", index: sizeLocal });
+          fctx.body.push({ op: "array.new_default", typeIdx: arrTypeIdx });
+          fctx.body.push({ op: "struct.new", typeIdx: vecTypeIdx });
+          return { kind: "ref_null", typeIdx: vecTypeIdx };
+        }
+
+        // new TypedArray(arrayLike) — copy from source array
+        // Compile source, then copy elements
+        const srcResult = compileExpression(ctx, fctx, args[0]!);
+        if (srcResult && (srcResult.kind === "ref" || srcResult.kind === "ref_null")) {
+          const srcTypeIdx = (srcResult as { typeIdx: number }).typeIdx;
+          const srcTypeDef = ctx.mod.types[srcTypeIdx];
+          // Check if source is a vec struct
+          if (
+            srcTypeDef?.kind === "struct" &&
+            srcTypeDef.fields[0]?.name === "length" &&
+            srcTypeDef.fields[1]?.name === "data"
+          ) {
+            const srcVecLocal = allocLocal(fctx, `__ta_src_${fctx.locals.length}`, srcResult);
+            fctx.body.push({ op: "local.set", index: srcVecLocal });
+            // Get source length
+            fctx.body.push({ op: "local.get", index: srcVecLocal });
+            fctx.body.push({
+              op: "struct.get",
+              typeIdx: srcTypeIdx,
+              fieldIdx: 0,
+            });
+            const lenLocal = allocLocal(fctx, `__ta_len_${fctx.locals.length}`, { kind: "i32" });
+            fctx.body.push({ op: "local.tee", index: lenLocal });
+            // Create new array of that length
+            fctx.body.push({ op: "local.get", index: lenLocal });
+            fctx.body.push({ op: "array.new_default", typeIdx: arrTypeIdx });
+            const dstDataLocal = allocLocal(fctx, `__ta_dst_${fctx.locals.length}`, {
+              kind: "ref",
+              typeIdx: arrTypeIdx,
+            });
+            fctx.body.push({ op: "local.set", index: dstDataLocal });
+
+            // If source and dest have the same array type, use array.copy
+            const srcArrTypeIdx = getArrTypeIdxFromVec(ctx, srcTypeIdx);
+            if (srcArrTypeIdx === arrTypeIdx) {
+              fctx.body.push({ op: "local.get", index: dstDataLocal });
+              fctx.body.push({ op: "i32.const", value: 0 });
+              fctx.body.push({ op: "local.get", index: srcVecLocal });
+              fctx.body.push({
+                op: "struct.get",
+                typeIdx: srcTypeIdx,
+                fieldIdx: 1,
+              });
+              fctx.body.push({ op: "i32.const", value: 0 });
+              fctx.body.push({ op: "local.get", index: lenLocal });
+              fctx.body.push({
+                op: "array.copy",
+                dstTypeIdx: arrTypeIdx,
+                srcTypeIdx: arrTypeIdx,
+              });
+            } else if (srcArrTypeIdx >= 0) {
+              const srcArrDef = ctx.mod.types[srcArrTypeIdx];
+              const dstArrDef = ctx.mod.types[arrTypeIdx];
+              if (srcArrDef?.kind === "array" && dstArrDef?.kind === "array") {
+                const srcDataLocal = allocLocal(fctx, `__ta_src_data_${fctx.locals.length}`, {
+                  kind: "ref",
+                  typeIdx: srcArrTypeIdx,
+                });
+                const copyIndexLocal = allocLocal(fctx, `__ta_copy_i_${fctx.locals.length}`, { kind: "i32" });
+                fctx.body.push({ op: "local.get", index: srcVecLocal });
+                fctx.body.push({ op: "struct.get", typeIdx: srcTypeIdx, fieldIdx: 1 });
+                fctx.body.push({ op: "local.set", index: srcDataLocal });
+                fctx.body.push({ op: "i32.const", value: 0 });
+                fctx.body.push({ op: "local.set", index: copyIndexLocal });
+
+                const srcGetOp =
+                  srcArrDef.element.kind === "i8"
+                    ? "array.get_u"
+                    : srcArrDef.element.kind === "i16"
+                      ? "array.get_s"
+                      : "array.get";
+                // (#2934 1c) The element-conversion matrix keys on the READ
+                // value's stack kind (packed i8/i16 widen to i32 via get_u/_s).
+                // The old matrix only knew f64↔int, so an EXTERNREF source
+                // element (`new Uint8Array([102])` where the literal compiled
+                // to an any[] externref vec) flowed uncoerced into the packed
+                // `array.set` — "array.set[2] expected i32, found array.get of
+                // externref" (the toBase64/`__cb_0` invalid-Wasm cluster). An
+                // externref element now unboxes (ToNumber) and truncates to
+                // integer storage; packed stores truncate width for free.
+                const srcReadKind =
+                  srcArrDef.element.kind === "i8" || srcArrDef.element.kind === "i16" ? "i32" : srcArrDef.element.kind;
+                const dstStoreKind =
+                  dstArrDef.element.kind === "i8" || dstArrDef.element.kind === "i16" ? "i32" : dstArrDef.element.kind;
+                let convertInstrs: Instr[];
+                if (srcReadKind === "externref" && dstStoreKind !== "externref") {
+                  // ToNumber the boxed element via the single coercion table
+                  // (#2108 — coercionPlan's externref→i32 row is exactly
+                  // unbox + trunc_sat; externref→f64 is the bare unbox).
+                  // Integer storage then truncates width on the packed store.
+                  addUnionImports(ctx);
+                  const plan = coercionPlan(
+                    { kind: "externref" },
+                    { kind: dstStoreKind === "f64" ? "f64" : "i32" },
+                    {
+                      boxNumberIdx: ctx.funcMap.get("__box_number") ?? null,
+                      unboxNumberIdx: ctx.funcMap.get("__unbox_number") ?? null,
+                    },
+                  );
+                  convertInstrs = plan?.instrs ?? [];
+                } else if (srcReadKind === "f64" && dstStoreKind !== "f64" && dstStoreKind !== "externref") {
+                  convertInstrs = [{ op: "i32.trunc_sat_f64_s" }];
+                } else if (srcReadKind !== "f64" && srcReadKind !== "externref" && dstStoreKind === "f64") {
+                  convertInstrs = [{ op: "f64.convert_i32_u" }];
+                } else {
+                  convertInstrs = [];
+                }
+
+                fctx.body.push({
+                  op: "block",
+                  blockType: { kind: "empty" },
+                  body: [
+                    {
+                      op: "loop",
+                      blockType: { kind: "empty" },
+                      body: [
+                        { op: "local.get", index: copyIndexLocal },
+                        { op: "local.get", index: lenLocal },
+                        { op: "i32.ge_u" },
+                        { op: "br_if", depth: 1 },
+                        { op: "local.get", index: dstDataLocal },
+                        { op: "local.get", index: copyIndexLocal },
+                        { op: "local.get", index: srcDataLocal },
+                        { op: "local.get", index: copyIndexLocal },
+                        { op: srcGetOp, typeIdx: srcArrTypeIdx },
+                        ...convertInstrs,
+                        { op: "array.set", typeIdx: arrTypeIdx },
+                        { op: "local.get", index: copyIndexLocal },
+                        { op: "i32.const", value: 1 },
+                        { op: "i32.add" },
+                        { op: "local.set", index: copyIndexLocal },
+                        { op: "br", depth: 0 },
+                      ],
+                    },
+                  ],
+                });
+              }
+            }
+            // Build result vec struct
+            fctx.body.push({ op: "local.get", index: lenLocal });
+            fctx.body.push({ op: "local.get", index: dstDataLocal });
+            fctx.body.push({ op: "struct.new", typeIdx: vecTypeIdx });
+            return { kind: "ref_null", typeIdx: vecTypeIdx };
+          }
+        }
+        // Fallback: treat argument as length
+        // (source was already compiled and is on stack — drop it and recompile as f64)
+        if (srcResult) fctx.body.push({ op: "drop" });
+        compileExpression(ctx, fctx, args[0]!, { kind: "f64" });
+        fctx.body.push({ op: "i32.trunc_sat_f64_s" });
+        const fallbackSize = allocLocal(fctx, `__ta_fsz_${fctx.locals.length}`, { kind: "i32" });
+        fctx.body.push({ op: "local.tee", index: fallbackSize });
+        fctx.body.push({ op: "local.get", index: fallbackSize });
+        fctx.body.push({ op: "array.new_default", typeIdx: arrTypeIdx });
+        fctx.body.push({ op: "struct.new", typeIdx: vecTypeIdx });
+        return { kind: "ref_null", typeIdx: vecTypeIdx };
+      }
+
+      // (#3054 B2) `new <TA>(buffer, byteOffset[, length])` — windowed
+      // shared-backing view. Build a `$__ta_view` with the byteOffset field
+      // populated (B1 pinned it 0) so `a[i]` addresses `byteOffset + i*width`
+      // into the SHARED buffer (sibling/DataView windows mutually observable).
+      // Standalone/WASI only — host-mode buffers are host objects, not native
+      // vecs (#1670). MUST match `inferTaViewType`'s multi-arg gate so the
+      // local's type and the constructed value agree.
+      if (noJsHost(ctx) && args.length >= 2 && !ts.isNumericLiteral(args[0]!)) {
+        // (#1930) Query the type-oracle boundary, not the raw checker — this
+        // also keeps the gate in lock-step with `inferTaViewType` (variables.ts),
+        // which resolves the buffer arg through the same oracle.
+        const winArgSymName = ctx.oracle.builtinReceiverOf(args[0]!);
+        if (winArgSymName === "ArrayBuffer" || winArgSymName === "SharedArrayBuffer" || winArgSymName === "DataView") {
+          const winResult = emitTaViewConstructWindowed(
+            ctx,
+            fctx,
+            args[0]!,
+            args[1]!,
+            args[2],
+            expr.expression.text,
+            (e, h) => compileExpression(ctx, fctx, e, h),
+          );
+          if (winResult) return winResult;
+        }
+      }
+
+      // new TypedArray() with multiple args — shouldn't happen per spec, but handle gracefully
+      // Treat like new TypedArray(0)
+      fctx.body.push({ op: "i32.const", value: 0 });
+      fctx.body.push({ op: "i32.const", value: 0 });
+      fctx.body.push({ op: "array.new_default", typeIdx: arrTypeIdx });
+      fctx.body.push({ op: "struct.new", typeIdx: vecTypeIdx });
+      return { kind: "ref_null", typeIdx: vecTypeIdx };
+    }
+  }
+  return NEW_GLOBAL_FALLTHROUGH;
+}
